split: extract AdminCredentialService and delete AdminService
Moves updateSmtpConfig, sendPasswordResetEmail, setTemporaryPassword,
resendVerificationEmail, unlockUser into a focused credential service.
AdminService is deleted entirely after the 4-way split. Shared AdminResult
and AdminError sealed types move to their own file. AdminSessionAuditRoutes
drops its unused adminService dependency.
