+++ src/main/kotlin/com/kauth/adapter/web/admin/AdminUserRoutes.kt
@@ -6,8 +6,8 @@ import com.kauth.domain.model.UserId
 import com.kauth.domain.port.AuditLogRepository
 import com.kauth.domain.port.SessionRepository
 import com.kauth.domain.port.UserRepository
+import com.kauth.domain.service.AdminCredentialService
 import com.kauth.domain.service.AdminResult
-import com.kauth.domain.service.AdminService
 import com.kauth.domain.service.AttributeResult
 import com.kauth.domain.service.ImpersonationService
 import com.kauth.domain.service.RoleGroupService
@@ -26,7 +26,7 @@ import io.ktor.server.routing.post
 import io.ktor.server.routing.route
 
 fun Route.adminUserRoutes(
-    adminService: AdminService,
+    credentialService: AdminCredentialService,
     adminUserService: com.kauth.domain.service.AdminUserService,
     roleGroupService: RoleGroupService,
     sessionRepository: SessionRepository,
@@ -254,7 +254,7 @@ fun Route.adminUserRoutes(
                         ?: return@post call.respond(HttpStatusCode.BadRequest)
                 val workspace = call.attributes[WorkspaceAttr]
                 val slug = workspace.slug
-                when (adminService.unlockUser(userId, workspace.id)) {
+                when (credentialService.unlockUser(userId, workspace.id)) {
                     is AdminResult.Success ->
                         call.respondRedirect("/admin/workspaces/$slug/users/${userId.value}?saved=unlocked")
                     is AdminResult.Failure ->
@@ -377,7 +377,7 @@ fun Route.adminUserRoutes(
                 val workspace = call.attributes[WorkspaceAttr]
                 val slug = workspace.slug
                 val baseUrl = call.resolvedBaseUrl()
-                when (adminService.resendVerificationEmail(userId, workspace.id, baseUrl)) {
+                when (credentialService.resendVerificationEmail(userId, workspace.id, baseUrl)) {
                     is AdminResult.Success ->
                         call.respondRedirect("/admin/workspaces/$slug/users/${userId.value}?saved=verification_sent")
                     is AdminResult.Failure ->
@@ -394,7 +394,7 @@ fun Route.adminUserRoutes(
                 val workspace = call.attributes[WorkspaceAttr]
                 val slug = workspace.slug
                 val baseUrl = call.resolvedBaseUrl()
-                when (adminService.sendPasswordResetEmail(userId, workspace.id, baseUrl)) {
+                when (credentialService.sendPasswordResetEmail(userId, workspace.id, baseUrl)) {
                     is AdminResult.Success ->
                         call.respondRedirect("/admin/workspaces/$slug/users/${userId.value}?saved=reset_email_sent")
                     is AdminResult.Failure ->
@@ -411,7 +411,7 @@ fun Route.adminUserRoutes(
                 val workspace = call.attributes[WorkspaceAttr]
                 val slug = workspace.slug
                 val baseUrl = call.resolvedBaseUrl()
-                when (val result = adminService.setTemporaryPassword(userId, workspace.id)) {
+                when (val result = credentialService.setTemporaryPassword(userId, workspace.id)) {
                     is AdminResult.Success -> {
                         val changeUrl = "$baseUrl/t/$slug/change-password?token=${result.value}"
                         val flashToken = FlashStore.put(changeUrl)
