+++ src/main/kotlin/com/kauth/adapter/web/api/ApiRoutes.kt
@@ -8,7 +8,7 @@ import com.kauth.domain.port.RateLimiterPort
 import com.kauth.domain.port.RoleRepository
 import com.kauth.domain.port.SessionRepository
 import com.kauth.domain.port.TenantRepository
-import com.kauth.domain.service.AdminService
+import com.kauth.domain.service.AdminCredentialService
 import com.kauth.domain.service.ApiKeyService
 import com.kauth.domain.service.CorsService
 import com.kauth.domain.service.EmailOtpService
@@ -37,7 +37,7 @@ fun Route.apiRoutes(
     sessionRepository: SessionRepository,
     auditLogRepository: AuditLogRepository,
     roleGroupService: RoleGroupService,
-    adminService: AdminService,
+    credentialService: AdminCredentialService,
     adminUserService: com.kauth.domain.service.AdminUserService,
     applicationManagementService: com.kauth.domain.service.ApplicationManagementService,
     userAttributeService: UserAttributeService,
@@ -114,7 +114,7 @@ fun Route.apiRoutes(
                 }
             install(apiContextPlugin)
 
-            apiUserRoutes(adminService, adminUserService, roleGroupService)
+            apiUserRoutes(credentialService, adminUserService, roleGroupService)
             apiRbacRoutes(roleRepository, groupRepository, roleGroupService)
             apiApplicationRoutes(applicationRepository, applicationManagementService, roleGroupService)
             apiSessionAuditRoutes(sessionRepository, auditLogRepository)
