+++ src/test/kotlin/com/kauth/adapter/web/admin/AdminUserAttributesAndClaimMappersTest.kt
@@ -8,7 +8,7 @@ import com.kauth.domain.model.TenantTheme
 import com.kauth.domain.model.User
 import com.kauth.domain.model.UserAttribute
 import com.kauth.domain.model.UserId
-import com.kauth.domain.service.AdminService
+import com.kauth.domain.service.AdminCredentialService
 import com.kauth.domain.service.RoleGroupService
 import com.kauth.domain.service.UserAttributeService
 import com.kauth.domain.service.UserSelfServiceService
@@ -354,7 +354,7 @@ class AdminUserAttributesAndClaimMappersTest {
         )
 
     private fun buildAdminService() =
-        AdminService(
+        AdminCredentialService(
             tenantRepository = tenantRepo,
             userRepository = userRepo,
             auditLog = auditLogPort,
@@ -418,7 +418,7 @@ class AdminUserAttributesAndClaimMappersTest {
                 call.respond(HttpStatusCode.OK, "session set")
             }
             adminRoutes(
-                adminService = buildAdminService(),
+                credentialService = buildAdminService(),
                 workspaceSettingsService =
                     com.kauth.domain.service
                         .WorkspaceSettingsService(tenantRepo, auditLogPort),
