+++ src/test/kotlin/com/kauth/adapter/web/api/ApiClaimMapperRoutesTest.kt
@@ -344,8 +344,8 @@ class ApiClaimMapperRoutesTest {
                         applicationRepository = com.kauth.fakes.FakeApplicationRepository(),
                         auditLog = com.kauth.fakes.FakeAuditLogPort(),
                     ),
-                adminService =
-                    com.kauth.domain.service.AdminService(
+                credentialService =
+                    com.kauth.domain.service.AdminCredentialService(
                         tenantRepository = tenantRepo,
                         userRepository = userRepo,
                         auditLog = com.kauth.fakes.FakeAuditLogPort(),
